requirements.txt
asttokens==3.0.0
colorama==0.4.6
executing==2.2.0
icecream==2.1.4
mypy==1.15.0
mypy_extensions==1.1.0
pillow==11.2.1
Pygments==2.19.1
typing_extensions==4.13.2
